garage-door: initial version of Go monitoring service
The Go monitoring service runs on the Raspberry Pi, receiving the
status of the garage door every 500ms. This service keeps track of the
door state. When it changes, it sends a request to an external web
endpoint. This endpoint happens to be hosted by AWS Lambda, storing
the state to S3, with another endpoint for the Amazon Alexa to query
this status.